+++ src/crypto/ossl/stream.rs
@@ -217,6 +217,7 @@ impl TryFrom<SrtpProfileId> for SrtpProfile {
         match value {
             SrtpProfileId::SRTP_AES128_CM_SHA1_80 => Ok(SrtpProfile::Aes128CmSha1_80),
             SrtpProfileId::SRTP_AEAD_AES_128_GCM => Ok(SrtpProfile::AeadAes128Gcm),
+            SrtpProfileId::SRTP_AEAD_AES_256_GCM => Ok(SrtpProfile::AeadAes256Gcm),
             x => Err(io::Error::new(
                 io::ErrorKind::Other,
                 format!("Unsupported SRTP profile {:x}", x.as_raw()),