+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
  * ice: fix panic on STUN Binding Indication #690
+ * bwe: fix panic on history monotonicity assert in `TrendlineEstimator` #700
 
 # 0.10.0